+++ Source/AtCoderLibrary/DataStructure/FenwickTree.cs
@@ -22,7 +22,7 @@ namespace AtCoder
     /// <typeparam name="TOp">配列要素の操作を表す型</typeparam>
     [DebuggerTypeProxy(typeof(FenwickTree<,>.DebugView))]
     public class FenwickTree<TValue, TOp>
-        where TOp : struct, IAdditionOperator<TValue>
+        where TOp : struct, IAdditionOperator<TValue>, ISubtractOperator<TValue>
     {
         private static readonly TOp op = default;