+++ Test/AtCoderLibrary.Test/Math/DynamicModIDTest.cs
@@ -1,0 +1,7 @@
+﻿namespace AtCoder
+{
+    public class DynamicModIDTest
+    {
+        internal struct DynamicModID : IDynamicModID { }
+    }
+}