+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.8.0] - Unreleased
-- Move operator interfaces to AtCoder.Operators namespace
+- **Breaking** Move operator interfaces to AtCoder.Operators namespace
+- Update CreateOperatorCodeFixProvider
 
 ## [1.7.0] - 2021-09-14
 - Raname PriorityQueue<TKey, TValue> to PriorityQueueDictionary<TKey, TValue>